refactor(auth): move WMS service auth under system domain
