--- app/service_auth/__init__.py
@@ -1,4 +1,0 @@
-# app/service_auth/__init__.py
-from __future__ import annotations
-
-"""WMS service-to-service authorization package."""
