+++ app/wms/system/service_auth/__init__.py
@@ -1,0 +1,4 @@
+# app/wms/system/service_auth/__init__.py
+from __future__ import annotations
+
+"""WMS system service-to-service authorization package."""
