+++ app/wms/system/service_auth/models/__init__.py
@@ -1,0 +1,20 @@
+# app/wms/system/service_auth/models/__init__.py
+from __future__ import annotations
+
+from app.wms.system.service_auth.models.wms_service_capability import (
+    WmsServiceCapability,
+)
+from app.wms.system.service_auth.models.wms_service_capability_route import (
+    WmsServiceCapabilityRoute,
+)
+from app.wms.system.service_auth.models.wms_service_client import WmsServiceClient
+from app.wms.system.service_auth.models.wms_service_permission import (
+    WmsServicePermission,
+)
+
+__all__ = [
+    "WmsServiceCapability",
+    "WmsServiceCapabilityRoute",
+    "WmsServiceClient",
+    "WmsServicePermission",
+]
