+++ app/wms/system/service_auth/repos/__init__.py
@@ -1,0 +1,10 @@
+# app/wms/system/service_auth/repos/__init__.py
+from __future__ import annotations
+
+from app.wms.system.service_auth.repos.wms_service_permission_repo import (
+    WmsServicePermissionRepo,
+)
+
+__all__ = [
+    "WmsServicePermissionRepo",
+]
