+++ app/wms/system/service_auth/services/__init__.py
@@ -1,0 +1,10 @@
+# app/wms/system/service_auth/services/__init__.py
+from __future__ import annotations
+
+from app.wms.system.service_auth.services.wms_service_permission_service import (
+    WmsServicePermissionService,
+)
+
+__all__ = [
+    "WmsServicePermissionService",
+]
