+++ tests/api/test_wms_read_v1_warehouses_api.py
@@ -6,7 +6,7 @@ from typing import Any
 from fastapi import APIRouter, FastAPI
 from fastapi.testclient import TestClient
 
-from app.service_auth.deps import (
+from app.wms.system.service_auth.deps import (
     WMS_SERVICE_CLIENT_HEADER,
     get_wms_service_permission_service,
 )
